requirements.txt
fastapi[standard]
python-dotenv
requests
redis